Update target frameworks and fix culture-related test issue
The target framework versions have been updated for EventHandler, ExtensionTest, and ExtensionBenchmark projects. netstandard2.1 and net8.0 are now the new target versions for Handlebars.Extension, Handlebars.Extension.Test, and Handlebars.Extension.Benchmark. The BenchmarkDotNet package is also updated from 0.12.1 to 0.13.12. In addition, a fix is applied in JsonTests to ensure proper functioning in different cultures by setting the current culture to invariant explicitly.